feat(pmproxy): /metrics endpoint (Prometheus text format)
Hand-rolled, no metrics-ecosystem dep so the Lambda zip stays small.
Atomic counters with one fetch_add per increment on the hot path.

Counter families:
- pmproxy_requests_total{route,status}   per-route request counts
- pmproxy_auth_failures_total{reason}    JWT validation failure breakdown
- pmproxy_rate_limit_drops_total         429 rejections
- pmproxy_jwks_refreshes_total{result}   ok / error
- pmproxy_ws_connections_total           lifetime WS upgrades
- pmproxy_ws_connections_active          gauge of open WS bridges
- pmproxy_tenants_active                 gauge from TenantRateLimiter

Wiring:
- ProxyState now carries Arc<Metrics>; always populated.
- authenticate() maps each AuthError → a stable failure reason.
- proxy_handler records (route_label, status_code) for every dispatch.
- JwksCache::new_with_metrics() optional sink; success/failure recorded.
- ws::ws_handler increments on upgrade, ws::handle_socket decrements
  when either tunnel half closes.

Exposed at GET /metrics, no auth (scrapers can't carry a Cognito JWT).
Content-Type per Prometheus exposition format 0.0.4.

13 → 20 lib tests (+5 upstream, +2 metrics). Both feature builds clean.

+++ pmtrader/cli.py
@@ -287,7 +287,7 @@ def _order_opts(f):
     f = click.option("--ttl", default=None, help="Auto-cancel after duration (e.g. '30m', '1h30m'). Needs running pmengine.")(f)
     f = click.option("--tick", default=None, help="Tick override; setting this skips engine routing.")(f)
     f = click.option("--price", default=None, type=float, help="Explicit limit price (omit for marketable sweep).")(f)
-    f = click.option("--size", default=None, type=int, help="Order size in shares.")(f)
+    f = click.option("--size", default=None, type=float, help="Order size in shares.")(f)
     f = click.option("--amount", default=None, help="Order size in USD notional (e.g. '$200').")(f)
     f = click.option("--match", default=None, help="Disambiguate sub-markets by keyword.")(f)
     return f
